+++ src/tests/portpicker_test.py
@@ -22,15 +22,19 @@ from contextlib import ExitStack
 import errno
 import os
 import socket
+import subprocess
 import sys
+import time
 import unittest
 from unittest import mock
 
 import portpicker
 _winapi = portpicker._winapi
 
+# pylint: disable=invalid-name,protected-access,missing-class-docstring,missing-function-docstring
 
-class PickUnusedPortTest(unittest.TestCase):
+
+class CommonTestMixin:
     def IsUnusedTCPPort(self, port):
         return self._bind(port, socket.SOCK_STREAM, socket.IPPROTO_TCP)
 
@@ -45,15 +49,62 @@ class PickUnusedPortTest(unittest.TestCase):
         portpicker._free_ports.clear()
         portpicker._random_ports.clear()
 
-    def testPickUnusedPortActuallyWorks(self):
-        """This test can be flaky."""
-        for _ in range(10):
-            port = portpicker.pick_unused_port()
-            self.assertTrue(self.IsUnusedTCPPort(port))
-            self.assertTrue(self.IsUnusedUDPPort(port))
 
-    @unittest.skipIf('PORTSERVER_ADDRESS' not in os.environ,
-                     'no port server to test against')
+@unittest.skipIf(
+        ('PORTSERVER_ADDRESS' not in os.environ) and
+        not hasattr(socket, 'AF_UNIX'),
+        'no existing port server; test launching code requires AF_UNIX.')
+class PickUnusedPortTestWithAPortServer(CommonTestMixin, unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        cls.portserver_process = None
+        if 'PORTSERVER_ADDRESS' not in os.environ:
+            # Launch a portserver child process for our tests to use if we are
+            # able to. Obviously not host-exclusive, but good for integration
+            # testing purposes on CI without a portserver of its own.
+            cls.portserver_address = '@pid%d-test-ports' % os.getpid()
+            try:
+                cls.portserver_process = subprocess.Popen(
+                        ['portserver.py',  # Installed in PATH within the venv.
+                         '--portserver_address=%s' % cls.portserver_address])
+            except EnvironmentError as err:
+                raise unittest.SkipTest(
+                        'Unable to launch portserver.py: %s' % err)
+            linux_addr = '\0' + cls.portserver_address[1:]  # The @ means 0.
+            # loop for a few seconds waiting for that socket to work.
+            err = '???'
+            for _ in range(123):
+                time.sleep(0.05)
+                try:
+                    ps_sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+                    ps_sock.connect(linux_addr)
+                except socket.error as err:  # pylint: disable=unused-variable
+                    continue
+                ps_sock.close()
+                break
+            else:
+                # The socket failed or never accepted connections, assume our
+                # portserver setup attempt failed and bail out.
+                if cls.portserver_process.poll() is not None:
+                    cls.portserver_process.kill()
+                    cls.portserver_process.wait()
+                cls.portserver_process = None
+                raise unittest.SkipTest(
+                        'Unable to connect to our own portserver.py: %s' % err)
+            # Point child processes at our shiny portserver process.
+            os.environ['PORTSERVER_ADDRESS'] = cls.portserver_address
+
+    @classmethod
+    def tearDownClass(cls):
+        if cls.portserver_process:
+            if os.environ.get('PORTSERVER_ADDRESS') == cls.portserver_address:
+                del os.environ['PORTSERVER_ADDRESS']
+            if cls.portserver_process.poll() is None:
+                cls.portserver_process.kill()
+                cls.portserver_process.wait()
+            cls.portserver_process = None
+
     def testPickUnusedCanSuccessfullyUsePortServer(self):
 
         with mock.patch.object(portpicker, '_pick_unused_port_without_server'):
@@ -67,8 +118,6 @@ class PickUnusedPortTest(unittest.TestCase):
             self.assertTrue(self.IsUnusedTCPPort(port))
             self.assertTrue(self.IsUnusedUDPPort(port))
 
-    @unittest.skipIf('PORTSERVER_ADDRESS' not in os.environ,
-                     'no port server to test against')
     def testPickUnusedCanSuccessfullyUsePortServerAddressKwarg(self):
 
         with mock.patch.object(portpicker, '_pick_unused_port_without_server'):
@@ -87,13 +136,21 @@ class PickUnusedPortTest(unittest.TestCase):
             finally:
                 os.environ['PORTSERVER_ADDRESS'] = addr
 
-    @unittest.skipIf('PORTSERVER_ADDRESS' not in os.environ,
-                     'no port server to test against')
     def testGetPortFromPortServer(self):
         """Exercise the get_port_from_port_server() helper function."""
         for _ in range(10):
             port = portpicker.get_port_from_port_server(
                 os.environ['PORTSERVER_ADDRESS'])
+            self.assertTrue(self.IsUnusedTCPPort(port))
+            self.assertTrue(self.IsUnusedUDPPort(port))
+
+
+class PickUnusedPortTest(CommonTestMixin, unittest.TestCase):
+
+    def testPickUnusedPortActuallyWorks(self):
+        """This test can be flaky."""
+        for _ in range(10):
+            port = portpicker.pick_unused_port()
             self.assertTrue(self.IsUnusedTCPPort(port))
             self.assertTrue(self.IsUnusedUDPPort(port))
 
@@ -375,6 +432,93 @@ class PickUnusedPortTest(unittest.TestCase):
         self.assertEqual(portpicker.pick_unused_port, portpicker.PickUnusedPort)
         self.assertEqual(portpicker.get_port_from_port_server,
                          portpicker.GetPortFromPortServer)
+
+
+def get_open_listen_tcp_ports():
+    netstat = subprocess.run(['netstat', '-lnt'], capture_output=True,
+                             encoding='utf-8')
+    if netstat.returncode != 0:
+        raise unittest.SkipTest('Unable to run netstat -lnt to list binds.')
+    rows = (line.split() for line in netstat.stdout.splitlines())
+    listen_addrs = (row[3] for row in rows if row[0].startswith('tcp'))
+    listen_ports = [int(addr.split(':')[-1]) for addr in listen_addrs]
+    return listen_ports
+
+
+@unittest.skipUnless((sys.executable and os.access(sys.executable, os.X_OK))
+                     or (os.environ.get('TEST_PORTPICKER_CLI') and
+                         os.access(os.environ['TEST_PORTPICKER_CLI'], os.X_OK)),
+                     'sys.executable portpicker.__file__ not launchable and '
+                     ' no TEST_PORTPICKER_CLI supplied.')
+class PortpickerCommandLineTests(unittest.TestCase):
+    def setUp(self):
+        self.main_py = portpicker.__file__
+
+    def _run_portpicker(self, pp_args, env_override=None):
+        env = dict(os.environ)
+        if env_override:
+            env.update(env_override)
+        if os.environ.get('TEST_PORTPICKER_CLI'):
+            pp_command = [os.environ['TEST_PORTPICKER_CLI']]
+        else:
+            pp_command = [sys.executable, '-m', 'portpicker']
+        return subprocess.run(pp_command + pp_args,
+                              capture_output=True,
+                              env=env,
+                              encoding='utf-8',
+                              check=False)
+
+    def test_command_line_help(self):
+        cmd = self._run_portpicker(['-h'])
+        self.assertNotEqual(0, cmd.returncode)
+        self.assertIn('usage', cmd.stdout)
+        self.assertIn('passed an arg', cmd.stdout)
+        cmd = self._run_portpicker(['--help'])
+        self.assertNotEqual(0, cmd.returncode)
+        self.assertIn('usage', cmd.stdout)
+        self.assertIn('passed an arg', cmd.stdout)
+
+    def test_command_line_help_text_dedented(self):
+        cmd = self._run_portpicker(['-h'])
+        self.assertNotEqual(0, cmd.returncode)
+        self.assertIn('\nIf passed an arg', cmd.stdout)
+        self.assertIn('\n  #!/bin/bash', cmd.stdout)
+        self.assertIn('\nOlder versions ', cmd.stdout)
+
+    def test_command_line_interface(self):
+        cmd = self._run_portpicker([str(os.getpid())])
+        cmd.check_returncode()
+        port = int(cmd.stdout)
+        self.assertNotEqual(0, port, msg=cmd)
+        listen_ports = sorted(get_open_listen_tcp_ports())
+        self.assertNotIn(port, listen_ports, msg='expected nothing to be bound to port.')
+
+    def test_command_line_interface_no_portserver(self):
+        cmd = self._run_portpicker([str(os.getpid())],
+                                   env_override={'PORTSERVER_ADDRESS': ''})
+        cmd.check_returncode()
+        port = int(cmd.stdout)
+        self.assertNotEqual(0, port, msg=cmd)
+        listen_ports = sorted(get_open_listen_tcp_ports())
+        self.assertNotIn(port, listen_ports, msg='expected nothing to be bound to port.')
+
+    def test_command_line_interface_no_portserver_bind_timeout(self):
+        # This test is timing sensitive and leaves that bind process hanging
+        # around consuming resources until it dies on its own unless the test
+        # runner kills the process group upon exit.
+        timeout = 9.5
+        before = time.monotonic()
+        cmd = self._run_portpicker([str(os.getpid()), str(timeout)],
+                                   env_override={'PORTSERVER_ADDRESS': ''})
+        self.assertEqual(0, cmd.returncode, msg=(cmd.stdout, cmd.stderr))
+        port = int(cmd.stdout)
+        self.assertNotEqual(0, port, msg=cmd)
+        if 'WARNING' in cmd.stderr:
+            raise unittest.SkipTest('bind timeout not supported on this platform.')
+        listen_ports = sorted(get_open_listen_tcp_ports())
+        self.assertIn(port, listen_ports, msg='expected port to be bound. '
+                      '%f seconds elapsed of %f bind timeout.' %
+                      (time.monotonic() - before, timeout))
 
 
 if __name__ == '__main__':